Swap tests: Abstraction & config
