+++ test/integration/swap/config.js
@@ -1,0 +1,18 @@
+export default {
+  bitcoin: {
+    rpc: {
+      host: 'http://localhost:18332',
+      username: 'bitcoin',
+      password: 'local321'
+    },
+    network: 'bitcoin_testnet'
+  },
+  ethereum: {
+    rpc: {
+      host: 'http://localhost:8545'
+    },
+    metaMaskConnector: {
+      port: 3333
+    }
+  }
+}
